refactor: split code into multiple submodules
